requirements.txt
Django==4.2.7
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
#Pillow==9.5.0  